::seeder error in order to generate JSON

+++ src/main/java/messages/VolAssociation.java
@@ -65,7 +65,7 @@ public class VolAssociation {
         this.capaciteLibre = capaciteLibre;
     }
 
-    public int getPrix() {
+    public long getPrix() {
         return prix;
     }
